feat: Migrate genomic coordinates from i32 to i64 for large genome support
- Vendor coitrees locally and upgrade Interval/IntervalNode fields from i32 to i64
- Fix nosimd feature flag in vendored coitrees lib.rs to always define COITree type alias
- Migrate all coordinate types across impg: QueryMetadata, SortedRanges, RawAlignmentInterval,
  DepthBfsHit, SerializableInterval, SubsettingResult, and all BFS/DFS function signatures
- Remove dead code revealed by pub(crate) visibility tightening in depth.rs
- Add audit report for depth command (notes/DEPTH_AUDIT_REPORT.md)

Note: existing .impgidx index files must be rebuilt due to changed binary layout.

+++ src/commands/lace.rs
@@ -1298,7 +1298,7 @@ fn create_gap_node<W: Write>(
 
     // Get gap sequence either from sequence index or create string of N's
     let gap_sequence = if let Some(index) = sequence_index {
-        match index.fetch_sequence(path_key, gap_start as i32, gap_end as i32) {
+        match index.fetch_sequence(path_key, gap_start as i64, gap_end as i64) {
             Ok(seq) => String::from_utf8(seq).unwrap_or_else(|e| {
                 error!("Failed to convert sequence to UTF-8: {e}");
                 "N".repeat(gap_size)
